+++ src/view/abstract.js
@@ -1,10 +1,13 @@
-import {createElement} from "../mock/utils";
+import {createElement} from "../utils/render.js";
 
 export default class Abstract {
   constructor() {
     if (new.target === Abstract) {
       throw new Error(`Can't instantiate Abstract, only concrete one.`);
     }
+
+    this._element = null;
+    this._callback = {};
   }
 
   getTemplate() {
