+++ src/view/films.js
@@ -1,4 +1,4 @@
-import AbstractView from "./abstract";
+import AbstractView from "./abstract.js";
 
 const createFilmsTemplate = () => {
   return `<section class="films">
