+++ src/view/profile.js
@@ -1,4 +1,4 @@
-import AbstractView from "./abstract";
+import AbstractView from "./abstract.js";
 
 const createProfileTemplate = () => {
   return `<section class="header__profile profile">
